pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "notebooklm2ppt"
version = "1.1.2"
description = "Công cụ chuyển đổi PDF thành bản trình chiếu PowerPoint có thể chỉnh sửa"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Tinh Nguyen", email = "tonyxbase@gmail.com"},
]
dependencies = [
    "PyMuPDF",
    "numpy",
    "requests",
    "spire.presentation",
    "python-pptx",
    "pywin32; sys_platform == 'win32'",
    "pywinauto; sys_platform == 'win32'",
    "scikit-image",
    "Pillow",
    "windnd; sys_platform == 'win32'",
]

[project.urls]
Homepage = "https://github.com/mrtinhnguyen/NotebookLM2PPT"
Repository = "https://github.com/mrtinhnguyen/NotebookLM2PPT.git"
Issues = "https://github.com/mrtinhnguyen/NotebookLM2PPT/issues"

[project.scripts]
notebooklm2ppt = "notebooklm2ppt.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["notebooklm2ppt"]
